pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "async_shutdown_test"
version = "0.1.0"
description = "Test project for async shutdown"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiokafka = {extras = ["snappy"], version = "^0.8.1"}
confluent-kafka = "^2.2.0"

[tool.poetry.dev-dependencies]
